// src/generated/model/favoriteModel.ts
/**
 * Datafood Server API
 * API contract of Datafood server application BRASIDAS
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FavoriteModel { 
    /**
     * ID of the favorite
     */
    id?: string;
    /**
     * ID of the product
     */
    productId?: string;
    /**
     * name of the product
     */
    name?: string;
    /**
     * name of the product
     */
    imageUrl?: string;
    /**
     * nutriscore of the product
     */
    nutriscore?: string;
    /**
     * ecoscore of the product
     */
    ecoscore?: string;
    /**
     * ID of the product
     */
    novagroup?: string;
    /**
     * Whether the favorite is active
     */
    isEnabled?: boolean;
    dateCreation?: string;
}